util: Avoid signed/unsigned integer comparisons
../../../lib/util/sys_rw.c: In function ‘sys_pread_full’:
../../../lib/util/sys_rw.c:219:25: warning: comparison of integer expressions of different signedness: ‘ssize_t’ {aka ‘long int’} and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
  219 |                 if (ret > curr_count) {
      |                         ^
../../../lib/util/sys_rw.c: In function ‘sys_pwrite_full’:
../../../lib/util/sys_rw.c:282:25: warning: comparison of integer expressions of different signedness: ‘ssize_t’ {aka ‘long int’} and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
  282 |                 if (ret > curr_count) {
      |                         ^
../../../lib/util/sys_rw.c: In function ‘sys_write_full’:
../../../lib/util/sys_rw.c:321:25: warning: comparison of integer expressions of different signedness: ‘ssize_t’ {aka ‘long int’} and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
  321 |                 if (ret > curr_count) {
      |                         ^

+++ lib/util/sys_rw.c
@@ -216,7 +216,7 @@ ssize_t sys_pread_full(int fd, void *buf, size_t count, off_t off)
 			break;
 		}
 
-		if (ret > curr_count) {
+		if ((size_t)ret > curr_count) {
 			errno = EIO;
 			return -1;
 		}
@@ -279,7 +279,7 @@ ssize_t sys_pwrite_full(int fd, const void *buf, size_t count, off_t off)
 			return -1;
 		}
 
-		if (ret > curr_count) {
+		if ((size_t)ret > curr_count) {
 			errno = EIO;
 			return -1;
 		}
@@ -318,7 +318,7 @@ ssize_t sys_write_full(int fd, const void *buf, size_t count)
 			return -1;
 		}
 
-		if (ret > curr_count) {
+		if ((size_t)ret > curr_count) {
 			errno = EIO;
 			return -1;
 		}
